Add copy connector step.
