+++ classes/local/step/connector_copy_file.php
@@ -1,5 +1,5 @@
 <?php
-// This file is part of Moodle - http://moodle.org/
+// This file is part of Moodle - https://moodle.org/
 //
 // Moodle is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -14,22 +14,24 @@
 // You should have received a copy of the GNU General Public License
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
+namespace tool_dataflows\local\step;
+
+use tool_dataflows\helper;
+
 /**
- * Version
+ * Copy file connector step
  *
- * @package   tool_dataflows
+ * @package    tool_dataflows
  * @author    Jason den Dulk <jasondendulk@catalyst-au.net>
- * @copyright  2022, Catalyst IT
+ * @copyright  2023, Catalyst IT
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
+class connector_copy_file extends connector_step {
+    use copy_file_trait;
 
-defined('MOODLE_INTERNAL') || die();
+    /** @var int[] number of output flows (min, max). */
+    protected $outputflows = [0, 1];
 
-$plugin->version = 2023050200;
-$plugin->release = 2022102600;
-$plugin->requires = 2017051500;    // Our lowest supported Moodle (3.3.0).
-$plugin->supported = [35, 401];    // Available as of Moodle 3.9.0 or later.
-// TODO $plugin->incompatible = ;  // Available as of Moodle 3.9.0 or later.
-$plugin->component = 'tool_dataflows';
-$plugin->maturity = MATURITY_ALPHA;
-$plugin->dependencies = [];
+    /** @var int[] number of output connectors (min, max). */
+    protected $outputconnectors = [0, 1];
+}
